+++ packages/session-replay/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * TBD
 
+## 1.1.0
+
+* Version bump in sync with `@splunk/otel-react-native`.
+* Bumped the native Splunk Android session replay SDK to `2.3.1`.
+
 ## 1.0.2
 
 * Version bump in sync with `@splunk/otel-react-native`.